LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-remove-src-expected.txt


FAIL invoking resource selection by setting src; await stable state assert_equals: expected 0 but got 3
